First pass at trying elements panel
