+++ packages/block-editor/src/hooks/color.js
@@ -8,8 +8,7 @@ import clsx from 'clsx';
  */
 import { addFilter } from '@wordpress/hooks';
 import { getBlockSupport } from '@wordpress/blocks';
-import { useMemo, Platform, useCallback } from '@wordpress/element';
-import { useSelect } from '@wordpress/data';
+import { useMemo, Platform } from '@wordpress/element';
 
 /**
  * Internal dependencies
@@ -19,20 +18,11 @@ import {
 	getColorObjectByAttributeValues,
 } from '../components/colors';
 import { __experimentalGetGradientClass } from '../components/gradients';
-import {
-	cleanEmptyObject,
-	transformStyles,
-	shouldSkipSerialization,
-} from './utils';
+import { transformStyles, shouldSkipSerialization } from './utils';
 import { getBackgroundImageClasses } from './background';
 import { useSettings } from '../components/use-settings';
 import InspectorControls from '../components/inspector-controls';
-import {
-	useHasColorPanel,
-	default as StylesColorPanel,
-} from '../components/global-styles/color-panel';
 import BlockColorContrastChecker from './contrast-checker';
-import { store as blockEditorStore } from '../store';
 
 export const COLOR_SUPPORT_KEY = 'color';
 
@@ -189,77 +179,6 @@ export function addSaveProps( props, blockNameOrType, attributes ) {
 	return props;
 }
 
-function styleToAttributes( style ) {
-	const textColorValue = style?.color?.text;
-	const textColorSlug = textColorValue?.startsWith( 'var:preset|color|' )
-		? textColorValue.substring( 'var:preset|color|'.length )
-		: undefined;
-	const backgroundColorValue = style?.color?.background;
-	const backgroundColorSlug = backgroundColorValue?.startsWith(
-		'var:preset|color|'
-	)
-		? backgroundColorValue.substring( 'var:preset|color|'.length )
-		: undefined;
-	const gradientValue = style?.color?.gradient;
-	const gradientSlug = gradientValue?.startsWith( 'var:preset|gradient|' )
-		? gradientValue.substring( 'var:preset|gradient|'.length )
-		: undefined;
-	const updatedStyle = { ...style };
-	updatedStyle.color = {
-		...updatedStyle.color,
-		text: textColorSlug ? undefined : textColorValue,
-		background: backgroundColorSlug ? undefined : backgroundColorValue,
-		gradient: gradientSlug ? undefined : gradientValue,
-	};
-	return {
-		style: cleanEmptyObject( updatedStyle ),
-		textColor: textColorSlug,
-		backgroundColor: backgroundColorSlug,
-		gradient: gradientSlug,
-	};
-}
-
-function attributesToStyle( attributes ) {
-	return {
-		...attributes.style,
-		color: {
-			...attributes.style?.color,
-			text: attributes.textColor
-				? 'var:preset|color|' + attributes.textColor
-				: attributes.style?.color?.text,
-			background: attributes.backgroundColor
-				? 'var:preset|color|' + attributes.backgroundColor
-				: attributes.style?.color?.background,
-			gradient: attributes.gradient
-				? 'var:preset|gradient|' + attributes.gradient
-				: attributes.style?.color?.gradient,
-		},
-	};
-}
-
-function ColorInspectorControl( { children, resetAllFilter } ) {
-	const attributesResetAllFilter = useCallback(
-		( attributes ) => {
-			const existingStyle = attributesToStyle( attributes );
-			const updatedStyle = resetAllFilter( existingStyle );
-			return {
-				...attributes,
-				...styleToAttributes( updatedStyle ),
-			};
-		},
-		[ resetAllFilter ]
-	);
-
-	return (
-		<InspectorControls
-			group="color"
-			resetAllFilter={ attributesResetAllFilter }
-		>
-			{ children }
-		</InspectorControls>
-	);
-}
-
 export function ContrastCheckerEdit( { clientId, name, settings } ) {
 	const enableContrastChecking =
 		Platform.OS === 'web' &&
@@ -278,85 +197,6 @@ export function ContrastCheckerEdit( { clientId, name, settings } ) {
 		<InspectorControls group="background">
 			<BlockColorContrastChecker clientId={ clientId } name={ name } />
 		</InspectorControls>
-	);
-}
-
-export function ColorEdit( {
-	clientId,
-	name,
-	setAttributes,
-	settings,
-	asWrapper,
-	label,
-	defaultControls,
-} ) {
-	const isEnabled = useHasColorPanel( settings );
-
-	const { style, textColor, backgroundColor, gradient } = useSelect(
-		( select ) => {
-			// Early return to avoid subscription when disabled
-			if ( ! isEnabled ) {
-				return {};
-			}
-			const {
-				style: _style,
-				textColor: _textColor,
-				backgroundColor: _backgroundColor,
-				gradient: _gradient,
-			} = select( blockEditorStore ).getBlockAttributes( clientId ) || {};
-			return {
-				style: _style,
-				textColor: _textColor,
-				backgroundColor: _backgroundColor,
-				gradient: _gradient,
-			};
-		},
-		[ clientId, isEnabled ]
-	);
-	const value = useMemo( () => {
-		return attributesToStyle( {
-			style,
-			textColor,
-			backgroundColor,
-			gradient,
-		} );
-	}, [ style, textColor, backgroundColor, gradient ] );
-
-	const onChange = ( newStyle ) => {
-		setAttributes( styleToAttributes( newStyle ) );
-	};
-
-	if ( ! isEnabled ) {
-		return null;
-	}
-
-	defaultControls = defaultControls
-		? defaultControls
-		: getBlockSupport( name, [
-				COLOR_SUPPORT_KEY,
-				'__experimentalDefaultControls',
-		  ] );
-
-	// Use provided wrapper or default to ColorInspectorControl
-	const Wrapper = asWrapper || ColorInspectorControl;
-
-	return (
-		<StylesColorPanel
-			as={ Wrapper }
-			panelId={ clientId }
-			settings={ settings }
-			value={ value }
-			onChange={ onChange }
-			defaultControls={ defaultControls }
-			label={ label }
-			enableContrastChecker={
-				false !==
-				getBlockSupport( name, [
-					COLOR_SUPPORT_KEY,
-					'enableContrastChecker',
-				] )
-			}
-		/>
 	);
 }
