+++ packages/block-editor/src/hooks/style.js
@@ -16,7 +16,8 @@ import { getCSSRules, compileCSS } from '@wordpress/style-engine';
  */
 import { BACKGROUND_SUPPORT_KEY, BackgroundImagePanel } from './background';
 import { BORDER_SUPPORT_KEY, BorderPanel, SHADOW_SUPPORT_KEY } from './border';
-import { COLOR_SUPPORT_KEY, ColorEdit, ContrastCheckerEdit } from './color';
+import { COLOR_SUPPORT_KEY, ContrastCheckerEdit } from './color';
+import { ElementsEdit } from './elements';
 import {
 	TypographyPanel,
 	TYPOGRAPHY_SUPPORT_KEY,
@@ -350,8 +351,8 @@ function BlockStyleControls( {
 	}
 	return (
 		<>
-			<ColorEdit { ...passedProps } />
 			<ContrastCheckerEdit { ...passedProps } />
+			<ElementsEdit { ...passedProps } />
 			<BackgroundImagePanel { ...passedProps } />
 			<TypographyPanel { ...passedProps } />
 			<BorderPanel { ...passedProps } />
