+++ packages/block-editor/src/style.scss
@@ -61,6 +61,7 @@
 @use "./hooks/block-bindings.scss" as *;
 @use "./hooks/border.scss" as *;
 @use "./hooks/color.scss" as *;
+@use "./hooks/elements.scss" as *;
 @use "./hooks/dimensions.scss" as *;
 @use "./hooks/layout.scss" as *;
 @use "./hooks/spacing.scss" as *;
